requirements.txt
azure-cognitiveservices-vision-face
flask